an implementation of a convolutional network
a low resolution version of the MNIST images that result from an
averaging a square of 4x4 pixels, thus transforming the 28x28 images
into 7x7 images. Implement this with a convolution operation and include
this part of your code in the solution submission. Compare the
performance to the previous performance.